+++ webtack/jsinterop-generator/src/test/fixtures/alt_inherit/output/gwt/main/java/com/example/Dictionary_requiredNullableBooleanSequenceValue.java
@@ -29,7 +29,7 @@ public interface Dictionary_requiredNullableBooleanSequenceValue {
   @JsOverlay
   @Nonnull
   static Builder requiredNullableBooleanSequenceValue(
-      @Nullable final Boolean[] requiredNullableBooleanSequenceValue) {
+      @Nullable final Boolean... requiredNullableBooleanSequenceValue) {
     return Js.<Builder>uncheckedCast( JsPropertyMap.of() ).requiredNullableBooleanSequenceValue( requiredNullableBooleanSequenceValue );
   }