+++ webtack/jsinterop-generator/src/test/fixtures/alt_inherit/output/gwt/main/java/com/example/Dictionary_requiredNullableShortSequenceValue.java
@@ -29,7 +29,7 @@ public interface Dictionary_requiredNullableShortSequenceValue {
   @JsOverlay
   @Nonnull
   static Builder requiredNullableShortSequenceValue(
-      @Nullable final double[] requiredNullableShortSequenceValue) {
+      @Nullable final double... requiredNullableShortSequenceValue) {
     return Js.<Builder>uncheckedCast( JsPropertyMap.of() ).requiredNullableShortSequenceValue( requiredNullableShortSequenceValue );
   }