+++ webtack/jsinterop-generator/src/test/fixtures/alt_inherit/output/j2cl/main/java/com/example/Dictionary_requiredAnySequenceValue.java
@@ -36,7 +36,7 @@ public interface Dictionary_requiredAnySequenceValue {
   static Builder create(@Nonnull final JsIterable<Any> requiredAnyIterableValue,
       @Nonnull final JsIteratorIterable<Any> requiredAnyIteratorIterableValue,
       @Nonnull final JsIterator<Any> requiredAnyIteratorValue,
-      @Nonnull final Any[] requiredAnySequenceValue) {
+      @Nonnull final Any... requiredAnySequenceValue) {
     return Js.<Builder>uncheckedCast( JsPropertyMap.of() ).requiredAnyIterableValue( requiredAnyIterableValue ).requiredAnyIteratorIterableValue( requiredAnyIteratorIterableValue ).requiredAnyIteratorValue( requiredAnyIteratorValue ).requiredAnySequenceValue( requiredAnySequenceValue );
   }