+++ webtack/jsinterop-generator/src/test/fixtures/alt_inherit/output/j2cl/main/java/com/example/Dictionary_requiredFloatSequenceValue.java
@@ -27,7 +27,7 @@ public interface Dictionary_requiredFloatSequenceValue {
 
   @JsOverlay
   @Nonnull
-  static Builder requiredFloatSequenceValue(@Nonnull final double[] requiredFloatSequenceValue) {
+  static Builder requiredFloatSequenceValue(@Nonnull final double... requiredFloatSequenceValue) {
     return Js.<Builder>uncheckedCast( JsPropertyMap.of() ).requiredFloatSequenceValue( requiredFloatSequenceValue );
   }