+++ webtack/jsinterop-generator/src/test/fixtures/alt_inherit/output/j2cl/main/java/com/example/Dictionary_requiredShortSequenceValue.java
@@ -27,7 +27,7 @@ public interface Dictionary_requiredShortSequenceValue {
 
   @JsOverlay
   @Nonnull
-  static Builder requiredShortSequenceValue(@Nonnull final double[] requiredShortSequenceValue) {
+  static Builder requiredShortSequenceValue(@Nonnull final double... requiredShortSequenceValue) {
     return Js.<Builder>uncheckedCast( JsPropertyMap.of() ).requiredShortSequenceValue( requiredShortSequenceValue );
   }