+++ webtack/jsinterop-generator/src/test/fixtures/alt_inherit/output/j2cl/main/java/com/example/Dictionary_requiredSomeTypeSequenceValue.java
@@ -28,7 +28,7 @@ public interface Dictionary_requiredSomeTypeSequenceValue {
   @JsOverlay
   @Nonnull
   static Builder requiredSomeTypeSequenceValue(
-      @Nonnull final SomeType[] requiredSomeTypeSequenceValue) {
+      @Nonnull final SomeType... requiredSomeTypeSequenceValue) {
     return Js.<Builder>uncheckedCast( JsPropertyMap.of() ).requiredSomeTypeSequenceValue( requiredSomeTypeSequenceValue );
   }