+++ webtack/jsinterop-generator/src/test/fixtures/dictionary/output/j2cl/main/java/com/example/GPURenderPassColorAttachment.java
@@ -1,0 +1,32 @@
+package com.example;
+
+import javax.annotation.Generated;
+import javax.annotation.Nonnull;
+import jsinterop.annotations.JsOverlay;
+import jsinterop.annotations.JsPackage;
+import jsinterop.annotations.JsType;
+import jsinterop.base.Js;
+import jsinterop.base.JsPropertyMap;
+
+@Generated("org.realityforge.webtack")
+@JsType(
+    isNative = true,
+    namespace = JsPackage.GLOBAL,
+    name = "GPURenderPassColorAttachment"
+)
+public interface GPURenderPassColorAttachment {
+  @JsOverlay
+  @Nonnull
+  static Builder create() {
+    return Js.uncheckedCast( JsPropertyMap.of() );
+  }
+
+  @Generated("org.realityforge.webtack")
+  @JsType(
+      isNative = true,
+      namespace = JsPackage.GLOBAL,
+      name = "GPURenderPassColorAttachment"
+  )
+  interface Builder extends GPURenderPassColorAttachment {
+  }
+}