+++ webtack/jsinterop-generator/src/test/fixtures/dictionary/output/j2cl/test/java/com/example/GPURenderPassColorAttachmentTestCompile.java
@@ -1,0 +1,12 @@
+package com.example;
+
+import javax.annotation.Generated;
+
+@Generated("org.realityforge.webtack")
+public final class GPURenderPassColorAttachmentTestCompile {
+  static GPURenderPassColorAttachment $typeReference$;
+
+  public static GPURenderPassColorAttachment.Builder create() {
+    return GPURenderPassColorAttachment.create();
+  }
+}