+++ webtack/jsinterop-generator/src/test/fixtures/member_types/output/j2cl/main/java/com/example/Dictionary_requiredNullableShortSequenceValue.java
@@ -28,7 +28,7 @@ public interface Dictionary_requiredNullableShortSequenceValue {
   @JsOverlay
   @Nonnull
   static Builder requiredNullableShortSequenceValue(
-      @Nullable final double[] requiredNullableShortSequenceValue) {
+      @Nullable final double... requiredNullableShortSequenceValue) {
     return Js.<Builder>uncheckedCast( JsPropertyMap.of() ).requiredNullableShortSequenceValue( requiredNullableShortSequenceValue );
   }