+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Convert the last parameter of a dictionary "create" method into a varargs method if it is an array type.
+
 ### [v0.25](https://github.com/akasha/akasha/tree/v0.25) (2021-09-26) · [Full Changelog](https://github.com/spritz/spritz/compare/v0.24...v0.25) · [API Differences](https://akasha.github.io/akasha-java/api-diff/?key=akasha-java&old=0.24&new=0.25) · [Source Diff](https://github.com/akasha/akasha-java/compare/v0.24...v0.25)
 
 The release includes 285 non breaking API changes, 6 potentially breaking API changes and 86 breaking API changes.